Refactor: condense MEModel dictlists
Currently, a MEModel object contains both a process_data
dictlist along with dictlists for each individual ProcessData
type. This is redundant and can lead to confusing behavior.
The individual ProcessData dictlists were removed and replaced
with getters that return a generator containing all instances
of the specific ProcessData type.

Updated JSONSCHEMA and json dumping and loading to reflect
this change.